requirements.txt
# ============================================================
# SmartVoyage 项目依赖
# ============================================================
# 安装命令: pip install -r requirements.txt
# Python版本: 3.9+
# ============================================================

# ------------------------------------------------------------
# 核心框架
# ------------------------------------------------------------
python-a2a>=0.1.0          # A2A协议SDK
mcp>=0.1.0                 # MCP协议SDK

# ------------------------------------------------------------
# LangChain生态
# ------------------------------------------------------------
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# ------------------------------------------------------------
# Web框架
# ------------------------------------------------------------
streamlit>=1.28.0          # Web界面

# ------------------------------------------------------------
# HTTP客户端
# ------------------------------------------------------------
httpx>=0.25.0              # 异步HTTP客户端
requests>=2.31.0           # 同步HTTP客户端

# ------------------------------------------------------------
# 数据库
# ------------------------------------------------------------
mysql-connector-python>=8.2.0

# ------------------------------------------------------------
# 工具库
# ------------------------------------------------------------
python-dotenv>=1.0.0       # 环境变量管理
pytz>=2023.3               # 时区处理
schedule>=1.2.0            # 定时任务
pydantic>=2.0.0            # 数据验证

# ------------------------------------------------------------
# 开发/测试工具
# ------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0              # 代码格式化
flake8>=6.1.0              # 代码检查
isort>=5.12.0              # import排序
mypy>=1.5.0                # 类型检查

# ------------------------------------------------------------
# 可选依赖
# ------------------------------------------------------------
# uvicorn>=0.23.0          # ASGI服务器
# gunicorn>=21.0.0         # WSGI服务器
# redis>=5.0.0             # Redis客户端(缓存)
